Store controller UUID for controller config in another table
There are parts of controller config that shouldn't be stored in the
controller config table as key and values. For example the controller
UUID should be part of a new controller table. The controller table can
then hold values that are known and unlikely to change, whilst also
adding the ability to add foreign key constraints to them (RI).

The service layer doesn't care how the data is laid out,
only that it can get a map of keys and values. The state layer
translates that to the right table locations.

We can use a view to simulate the old controller config table with
all the values, yet the controller config table just holds scalar
or simple string values.