Update container architecture on machine import.
Fixing a bug where containers archicture has not been saved previously.
Need to determine the charm platform, architecture/os/channel, when
using machine scoped placements directives in deploy.

Assume that the container has the same architecture as the host machine.
This is hardcoded for KVM containers today and what is supported for
LXC.