Removes Syntax sugar from core logger interface
The logger interface was copied from loggo.Logger type. It has a lot
of syntax sugar for accessing log levels, but in reality you can just
use IsLevelEnabled. We should just have one and exactly one way of doing
this not multiple. We're doing this to make it easier for our future
selves.