Improve run_deploy_charm_placement_directive.
Add a container to the mix due to LP:2064174. Due to issues with nested
LXD containers, use a virtual-machine for the host machine when testing
against the lxd provider only.

Added a check to see that the machine's base was used to deploy the
charm.

Implement a wait_for_container_agent_status for use in the updated test.