Rename of model finaliser to activator.
Naming things can be hard. Several months ago we introduced the concept
of finalising a model during the creation process. This allows a model
and all of the model components to be built up over disparate
transactions.

We use the keyworld finalised to represent a model and it's finished
state during this process. It has been decided to rename finalised to
activated better representing what is being done.