Export all LoginResultParams
Due to a mistake in the initial implementation of LoginProviders, the fields were left unexported. This disallows the capability of creating your own LoginProvider to Juju. The specific case here is a JWT login provider which is specific to JIMM and as such we wish to implement the provider in JIMM.