Cut the juju version
Remove the non numeric (tag) part of the juju version when setting
$JUJU_AGENT_VERSION. There is currently a terraform test which fails
becuase it tried to parse the version but only looks for numbers of the
form "X.Y.Z" and it errors on anything else.

We intend to update the terraform test to stop it relying on
this enviromental variable and determining what is supported another
way.