Assert change stream becomes idle
Adds a test function to ensure that the change stream becomes idle
before checking the side effect of the function. This ensures that
waiting for a certain set of watcher changes happen after a known
clean slate.

This should improve reliability of the watcher tests after a known
side effect.