Reworks model creation id generation.
When creating models in DQlite via the service we use a
ModelCreationArgs struct that drives all of the meta data associated with
the model. Via previous changes we were putting the responsibility back
on to the caller of the create model process to generate model id.

This is an anti pattern that we want to remove entirely as it should the
services layer generating this and returning it to the caller. This
change the caller now receives the model id back as a return value.

There still exists two cases where the model id will be supplied
upfront. The first is via model migration where we are importing a model
that already exists and so we must use the id that has already been
established. The second is during a controller bootstrap when making the
controller model. This second case we can eventually get rid of and move
to the same pattern as the rest of the call sites but it will require
some big re-work to the Juju client at a later stage.

This commit also makes changes to secret back end selection for newly
created models. In previous commits a decision was made to pick a models
"default" secret back end in the sate layer. There was no reason to have
done this as all the context required is at the services layer. The code
has been moved one level up and removes the need to do more computation
inside of a transaction.

Renamed owner permission function for new models so that it now reflects
intent correctly. That is when creating a new model we wish to grant the
owner of the model admin permissions.