Intern the juju/charm package
Interns the juju/charm package so that its dependencys are managed
directly by the juju/juju package. Prior to this change, each major
version of juju required changes to the charm package. Releasing a
new version of the charm package is time consuming and hard to get
right when an older major version has a bug with the assoicated
charm package. Rev'ing the package in the correct way is time
consuming. This should reduce the complexity of juju/juju where the
core charm logic is outsourced to another package.

Concretely the all code now points to the internal/charm package. We
will have to spend time slowly merging bother core/charm and
internal/charm. External clients will probably want access to this,
but we should resist that and expose interfaces for reading charms
and bundles. Complex reading of charms or bundles should be then
interned to the api itself (potentially with vistors or adapters).

Lastly, there is a package of charms that are currently living in
internal/charm/internal/test-charm-repo, those should be moved to
testcharms/charm-repo.