Remove series from charm URL and bundle data
Remove references to series in internal charm code related to charm url and
bundles

For the case of charm urls, to ensure we can still parse existing charm urls,
we do not remove the series attribute as such, but instead make it private.
It is then only accessible via stringifying a url. This ensures that parsing
then stringifying a charm url does not change it's value.

For the time being keep references to series in charm metadata. Wait until we
remove support for this in Juju itself