Merge common/imagecommon pkg into client/imagemetadata
This package was pulled out in #7631 to create a new controller-side image
metadata facade (#7636) which was later removed in #8400. Since this code is
only used by the client facade, we can take it out of common and into the
client facade package.

Re: tests, only TestSaveModelCfgFailed had to be moved over. The other tests
were already covered by existing tests in the client package.