Intern featureflag package
The featureflag package is a simple package that we have as a
dependency. The code in the featureflag package didn't make sense
in terms of the windows build, so I've drastically reduced the scope
of the package when I brought it into juju.

This is one less dependency to manage.