Drop support for series in charm meta
Also, at the same time, drop support for charms without manifests.

All charms in charmhub already have a manifest.yaml file. So this drop
in support only affects a small number of local charms, those which have
not been build with charmcraft

So this change can be viewed as a part of the broader drop in support
for charms not build with charmcraft. NOTE: You can still unpack charms
built by charmcraft and deploy them from dirs

We also remove support for Deployments in metadata. These were only used
for podspec charms, which are also no longer supported in juju 4.0

This also means we can drop GetBaseFromSeries!

In doing so, we need to fix a bunch of our test charms, which are now
invalid since they have no manifests