Watch cloud test fixes
The watch cloud test wasn't working how it was expected. The upsert
was just triggering a new change in the old code, prior to this patch.
The cloud config is currently immutable, I'll open a bug to fix this.
As the config is not updated and the trigger prevents a noop change
then the test fails. The solution is to then update a field on the
cloud that will actually cause a change. In this case the endpoint
being updated.