feat: Allows auto updating of LXD images
This is the first step to removing auto-upgrade from juju when
bootstrapping. We only do this during bootstrap and never any other
time. We do want to continue with updates, otherwise we won't be
able to install packages correctly. LXD provides ways to ensure
that auto update works ensuring better security around images.

To get this feature working we need to replace the fingerprint of
the image with the alias. This was copied from the LXD source code
which does the same thing[1]. There is some trickery ensuring that
we set the fingerprint back to what it was originally once we've
copied the remote image and setup the auto-update. The bit that
seems a bit brutish is that we have to splat the aliases to ensure
that they correctly line up with the fingerprints. This again
was taken from the lxd source code[2]. We're currently not able
to use the code from their source as it's not available.

 1. https://github.com/canonical/lxd/blob/f2dded07c9c4c046d3484f50c8248171b65352b6/lxc/image.go#L255-L256
 2. https://github.com/canonical/lxd/blob/f2dded07c9c4c046d3484f50c8248171b65352b6/lxc/utils.go#L181-L233