Remove SuperSubnets from environs networking
As part of dropping the support for fan networks in juju 4.0, we are as
a first step removing the SuperSubnets function from the environs
networking interface and the implementations in all providers.

One step further was taken as well, by removing the fan discovery in
container networking auto configure (in legacy state).
This means that now the container networking config can either be
"provider" or "local". Also, this method will be moved to the networking
domain in a future patch.