Fix model_config watch
The model config watch wasn't correct once the generated triggers
were added. Firstly, we deleted all of the keys and then re-added
them again even if they where the same. This caused a lot of churn
in the changestream. The fix was just ensure that we inserted,
updated and deleted rows correctly.

The second problem was that we would see the initial changes from
bootstrap and then see them again when the change stream was processing
requests. This might be ok for some watchers (I highly doubt that),
but the concept of watchers are, watch changes and then react. For
most things that take information from the API server, the initial
event is thrown away, so it's pointless trying to get that. It's
better to just read the state of the world then respond to changes.
The model_config watcher will now emit an empty initial event and
then subsequent events will tell you what has changed.