Rework change stream idle
We need to ensure that the change stream is actually idle. We
can do this by checking what the current upper bound is vs what is
the MAX change log id. If those match we're done!

This is only ever run in test code as we don't care for production.