feat: Rework fingerprint decision tree
Instead of passing a fingerprint as an argument and optionally set
it, we instead just set it on the sourced image and then reset it
always. This makes for a cleaner implementation.

In order to reason about the code and better understand what the
image alias is actually doing, I've cleaned up the nested for loop
from the image alias look up. This removes the potential quadratic
behaviour from alias look up.

Lastly, althoug ensureImageAliases was taken from the LXD source
code, we're now using the set.Strings type to find the intersection
for more efficent look ups.