External controller log message
Whilst debugging a production issue, we saw the log line that
controller info was updated, but we don't know what it went from
and to. As this happened a lot, it would be helpful to better
understand why we're constantly getting a new set of addresses.

Neither the cert or the alias are updated when we get new information,
so it doesn't make sense to keep spitting them out. Just the addresses
are enough to get an idea of the churn.