DDL SQL files
In order to get better IDE support for sql files, we can embed the
sql files into the binary directly. This provides best of both
worlds. We can have independant sql files, that we can validate,
format and test. All whilst we read from embedded strings. There
will be no disk reads when using this.

The files are read in order, so 0000-9999. This is per model, and
controllers will have their own independant range.

This only tackles model sql at the moment.