Remove support for fan subnets on the network domain
As part of the drop of fan networking support on juju 4.0, we are now
removing the modelling of fan networks on the new network domain.

This removes the subnet association tables and the subnet type tables
along with their references and all queries that used them. From this
patch forward, only (flat hierarchy) subnets are stored without any association between
them.