Charm and normalized charm metadata schema
Adds the charm and charm metadata schema in normalized form. As the
charm metadata has nested types this needs to exploded into indvidual
tables.

The only thing that hasn't been migrated across is the assumes field,
as this is self referential and modelling that might be tricky.

Charm manifest, config and actions will follow on. This PR is difficult
enough to grok with the explosion of tables.