Intern packaging package
The packaging package hasn't been maintained and the abstraction isn't
really holding it weight. In light of trying to own our dependencies
and understand what we're installing at what stage, we should bring
this in as some global constants don't make sense.

For instance the UbuntuCloudArchiveSigningKey isn't actually used
anywhere in the code base and we should probably remove it.

Additionally, do we need the list of cloud archive packages set directly
in the code. This feels like there were grand plans for this, but
was never fully implemented. As we move forward we could start removing
parts of the abstraction that don't make sense.

There are packages in the juju namespace that have been slowly rotting,
this is one of them.