No longer error out for series in bundles
After some further discussion on dropping support for series in bundles,
we have decided that erroring out is not the best approach.

This was triggered by realisation of the fact that charmhub still
enforces that uploaded bundles contain a series (or 'bundle', but that's
only relevant to k8s bundles).

Meaning it is guarenteed that (almost) all bundles in charmhub today would
break.

As such, instead of erroring out, we show a different warning stating
that the series key is unsuppoirted and will be ignored

The suggested solution to bundle authors to ensure they bundles are
supported in 4.0 is to use 'default-base' or 'base' in their bundles
instead.

However, 2.9 does not recognise default-base or base. So series needs to
be used to be supported by 2.9. As such, we need to support bundles with
both series and base provided, as we do in 3.x.

So we will only display this warning if a series is found without a
base, since this is the only scenerio which we have dropped support
for.