Remove UbuntuVersions function
This function is one of the few left that uses distro-info to determine
it's output.

This one, however, returns series and bases which both are and are not
supported. We wish to replace this whole system with just a hardcoded
list of supported bases. As such, this function needs to go

Fortunately, it was only used in 1 place. upgradevalidation. We used
this function to get a list of all unsupported ubuntu versions, which we
would then use to query state to check if any machines are deployed to
those bases.

Instead, we use WorkloadBases to get the supported bases, and then
assert that all machines are deployed to one of these bases.

In doing so, add "AllMachinesCount" state method, which returns an
integer. This isn't ideal, since we should be trying to avoid making
changes to state whilst it's being migrated to DQLite, but it turns out
the simpler signature removes a lot more complexity elsewhere. We no
longer need to shim out machines in our facades, and transform between
slices of identical interfaces, etc.