Lint all sql files
This formats and lints all the SQL files. We're using sqlfluff for
linting and formatting. The style isn't 100% nice, but then again
any formatter isn't perfect (see gofmt as an example).

The files are 100% consistent though.

In addition to formatting I found a lot of stupid and simple mistakes
that the linter found.

 - using key words for column names
 - missing commas
 - ambigous column names in joins
 - duplicated columns in views

I'm not even sure how some of the foreign keys worked (unless the sqlite
parser is very lenient) because missing comma where present in mutliple
locations.