Use's DQlite model creation id to drive model id.
While we are transitioning to DQlite we are currently writing to both
DQlite and Mongo for newly created models. Both these processes have
generation logic for the model id and can become in conflict. This
change takes the model id generated by DQlite and uses it as the model
if for Mongo.