Hardcode ControllerBases + WorkloadBases
It was decided in Madrid that, going forward, we want to provide a hard
coded list of ubuntu bases which we wish to support for
1) controllers
2) other workloads

So hardcode the output to ControllerBases & WorkloadBases

This means we can simplify the signature of these functions. This
resulted in many small changes across Juju

Also, importantly, this means we can remove all the code which searches
distro info. This simplifies core/base dramatically!

I also took the decision to remove
1) AllWorkloadVersions. This can simply be calculated using
   WorkloadBases.
2) AllWorkloadOSypes. This is only ubuntu now, so I just use the const instead