Add upgrade validation for fan networking model config
On juju 4.0 the Fan networking method is no longer supported, and along
with dropping it, some facades were dropped. This breaks model migration
and upgrades.

To fix this, we have decided to prevent the migration/upgrade of models
with the config container-networking-method=fan *when* there are
containers deployed. This is done by adding a new validation method on
model migration/upgrade.