Fix Azure provided image selection for daily stream
Also, simplify the logic somewhat.

We were incorrectly searching in the same Azure offer for mainline LTS
releases, and daily releases.

CPC have been releasing images the images we want for the daily stream
to a different offer, ending "-daily". Make sure we use the correct
stream!

Also, we do not include the release date in SKUs. In fact, each offer
only has a small, fixed numebr os SKUs. All that we need to do is select
the correct SKU, and that will give us the most recent image from that
channel.