Asserts controller config defaults to model.
When bootstrapping model config we were not applying the built in
defaults for model config to the model. Fortunately we didn't notice
because currently the client generates the defaults during bootstrap and
passes them to the controller.

However for tests using the bootstrap methods the defaults were not
being applied. This change now forces the bootstrap method to always
apply the model defaults on set.