Introduces helper for getting controller model.
Juju installs itself into a well known model at bootstrap time called
the "controller" model. The uuid for this model is unknown and changes
from controller to controller. There exists cases where we need to
extract this model for use. The only way to reliably do this is to find
the model based on the well known name and user that owns the model.

This commit introduces model service helper for retrieving the
controller model and hiding the specifics of how this is achieved from
the caller.