Return a non-empty network topology when only alpha space has been explicitly set as constraint.
At the provisioning info facade, we fill the ProvisioningNetworkTopology field of the provisiong info struct with the subnets mapped to the AZs. We have this information because we have the constraints (and enpdoint bindings) and we then do a lookup to find the subnets that are in that space(s).

We merge the constraints *and* the bound endpoint spaces (which by the way is `alpha` by default) and then return the list of spaces merged.
The issue is that later we make the following check: if only the alpha space is left after this merge, then we don't bother to fill a network topology that would have to be interpreted by the providers.
This check is wrong because it assumes that if we only have alpha in that resulting list it's because it was added by the default endpoint binding space. Another possibility is that the user had explicitly passed only the alpha space as a positive constraint.

This patch fixes this assumption, by not only checking if the merged list contains only alpha but also if the constraints weren't explicitly set to alpha only. If that's the case then we set a network topology and providers will be able to correctly instantiate the machine on the correct spaces (and in the case of openstack by creating only the correct ports).