Denote noble as a supported base
Our system to automatically pick up when Ubuntu versions become
unsupported doesn't make unsupported bases supported.

We haven't noticed this yet since we makred noble as ESMSupported, which
usually we '||' with Supported. However, this is not the case always
which leads to some usual bugs

i.e. upgrade-controller thinks noble is deprecated