fix: unit agent lost on migration success
During the migration minion's SUCCESS phase, at the
point of no return, the api connection can fail.
This leads to the doSUCCESS handler being unable to
report the SUCCESS phase as being successful to the
source controller. When this errors, the target IPs
and CA certs are not saved to the agent config.

This issue was seen with unit agents failing to use
the target controller, but the machine agent using
the target controller succesfully. This would show
the unit agent on the machine as lost.

This fix attempts to make the reporting back to the
source controller more robust to connection issues.
If for some reason, we fail these retry attempts, we
now have a very angry critical log message to inform
the admin of this situation.