Improve storage directive parsing
Our storage directive parsing is fairly naive. We just split on commas,
then iterative over the parts, filling in the parts of a Constraints
struct as we go

We check if the pool name is overwritten, returning an error if it is.
Do the same for size and count. This will mean that cases like "ebs,1024,3"
will fail, instead of passing silently