fix: panic: InsecureKeyProfile cannot be used outside tests
The modeloperator on k8s was panicking due to test code being
imported into the domain/secret/service package.

Firstly this adds a linker failure to prevent the testing package
from being linked into the juju binaries.

Secondly this adds a TODO watcher, which is useful in replacing the
abuse of the watchertest code.