fix(migration): Return ErrDBNotFound for not found namespace
If the namespace isn't found we should return a sentinel error, so
we prevent constant restarting. If we get into a restarting loop
we can't progress the model and we can't rollback.

This reinstates the readonly model rollback. The delete db should
only be used sparingly and ideally we should not require it. The
fallback is there as a safety net.

The error for agent version is still broken, but it forces the problem
to exist.