fix(offer): improve error message when updating offers with removed endpoints
Juju offer is an upsert operation but this might misleading to users.
Since we are not changing this behavior right now, this patch at least
improves the error message when *updating* an offer and in this update
an endpoint (with active consumers) is removed.
The new error message should give the user the hint that they are
actually updating the offer and therefore if they don't want to remove
the previously added endpoints, they should be added as a list in the
passed arguments. This behavior is already documented.

[squash] fix typo on ensureEndpointsNotInUse args

Co-authored-by: Simon Richardson <stickupkid@gmail.com>

[squash] add test with multiple endpoints offer update

[squash] fix cmd status tests