feat(sql): STRICT sql tables
SQLite by default uses dynamic untyped columns. There is no validation
of columns on the way in or out. The proposal is to enable strict
tables which helps with locating potential failures with DDL and
statement queries.

All DATETIME and TIMESTAMP columns become TEXT. Then BOOLEAN becomes
INT (represented as 0 or 1).