Turns off agent version checking for model import.
When importing models we were checking the fact that the agent version
was the same as the current juju version in use. The logic that we
currently mainting in Mongo is that we accept whatever version is set on
the imported model and let it go through. This is because as part of the
model migration process the agent binaries should be moved from one
controller to the other.

In an ideal world we would like to check that we have tools for the
model being imported but need to fully implement find tools in DQlite
before this can be done.

This commit addresses current issues where we can not import from 3.6 to
4.0.