feat(charm): Charm SQL DDL Schema
This is the follow on from the charm meta SQL DDL schema. This
includes, actions, config and manifests.

As actions params can be self referential, we're not attempting to
split the object into separate columns, instead it makes sense to
store them as just a blob of text. The potential for breakages is
high.

Charm config is normalized as a name, type with the default value.
The type is an enum of potential values; string, int, float, boolean
and secret. No validation of the content of the default value is
done.

Manifest table holds just the bases. There is a risk that the manifest
will change (it's an external data source, that isn't under the control
of juju), so keeping only a very small subset of the data is preferable.
We can add columns as we see fit.

With the addition of the manifest table, charm_channel table was
dropped. It's much better to explode the channel into it's constituent
parts (track, risk and branch) to ensure that normalisation happens
and data integrity is kept. The consequence of this, is that the
manifest table has a primary key that can contain NULL values. Sqlite
allows this. In addition, we never expect querying of a manifest base
individually.