Implements a ssh public key importer.
This commit replaces the need to the use
https://github.com/dustinkirkland/ssh-import-id/tree/master in Juju via
exec in the facades.

Currently this is not in use by Juju. This is the first pass at
implementation. We want to do this work so that we can avoid execing out
to the shell to run another program that could provide non deterministic
results.

This also allows us to use http clients that are configured for the http
proxy set in Juju.