fix(upgrade): noble upgrade support
The default base was selected by first checking for the latest LTS,
then, if that didn't work, trying the offical juju defualt.

This commit swaps the order of those, so that the officail juju default
version is preferd over the latest LTS. The latest LTS is still used if
the charm does not support the juju default version.