fix: stop setting a different LatestLTS to test.
There was a bug in this test where the "latest" variable used in the
table driven tests was set to LastestLTS before the mock value was set.
This meant it was 24.04 not the mocked value of 20.04. This caused the
tests to fail.

By removing this mocking we avoid this problem. The latest value is the
same globally and locally