style(sqlair): use SQLair everywhere in externalcontroller.
As part of the effort to move from using the database/sql library to
SQLair, get rid of all uses of StdTxn.

A todo to use IN, which SQLair now supports is implemented.

The NewUpdateStatements function is removed and the prepare functions
are moved back to where the queries are run. This function was added to
avoid redoing sqlair.Prepare every time we queried but this problem has
now been overcome with the use of the State.Prepare function.