fix(deploy): Pre-deployed charms no longer always deploy to series from curl
Instead we ignore the curl series field we construct a base selector.
This removes the one final read of charm url series in 3.6

This will need refactoring later. Using a base selector here duplicates
a lot of code. The regular local charms we're used to construct and
identical base selector in 'cmd/juju/application/deployer/deployer.go'
function 'determineBaseForLocalCharm'. We should refactor our factory
and deployer kinds to eliminate this duplicated code.

However, this will be much much easier after a backport. And this
backport can only be done after we remove this final read of series
(since it removes this series). So I'm not worrying about the hacking
way of doing this at this time. It will be improved in a subsequant
commit.