style(sqlair): in the objectstore and move UUID gen to state
The type objectstore.Metadata was local to the objectstore domain. With
the transition to SQLair I required a very similar type just in the
state with the db tags on it. For this reason I have moved the type to
state, tagged it, and replaced it in the service with the core object
store metadata type.

The UUID generation was done in the service layer but now can be done
simply in the state layer.