refactor(deploy): Remove duplicated base selector code
We use a base selector to determine the base to deploy a local charm to
in the factory before contructing the DeployerKind.

Do the same in the same place for pre deployed local charms.

A result of this, however, is that we need to add another method to the
api interface on GetDeployer. This resulted in a bit of refactoring

But overall, this cleaned up a lot of things around deploy code