refactor(deploy): Remove validateCharmBaseWithName
validateCharmBaseWithName was originally intended to validate a user
provided OS is supported by a charm when deploying

However, for a while now we have used a BaseSelector, which guarentees a
base which is supported by this charm. This validation is now
superfluous, and duplicates code in the BaseSelector.

So remove it

See discussion here:
https://github.com/juju/juju/pull/16101#discussion_r1301431892

Fixes:
https://bugs.launchpad.net/juju/+bug/2032639