feat: unit schema
Adds the rest of the unit schema. This includes adding foreign
keys to machine and charm. The weird thing is that machine has a
foreign key to the net_node along with the unit. I'm unsure if they
can diverge. If they can't, a foreign key to the machine should allow
the look up of the net node without storing the fk directly.

Other descisions:

 1. Tools metadata only stores the version. The charm metadata uuid
    is a fk to the object_store metadata, which holds the information
    with the hash and size.

 2. The unit config was added as a normalized exploded table.

The unit state was already added, so no changes are required.