feat: remove all watcher from schema.json
As we're no longer offering the all watcher and model watcher from
4.0 we need to stop advertising it. This will break a lot of clients
(pylibjuju and jslibjuju), but we won't be broadcasting delta change
events from juju any more.

A new implementation and spec are being actively worked on. We're
removing it early, so that we don't have to wire up the new domains
(charm, application, unit and machine) to something that we're removing.