feat: remove all watcher from bundle handler
Removes the all watcher from the bundle handler. We'll need to
circle back around to implement a different way of getting the machine
ID for the unit name. For this patch I've implemented a TODO.

As we're removing the watcher and all the params, so the wait-for
command also needs to go. This is 100% based on the all watcher. It
does not work at all without it. Removal of it for now and we can
add it back when we have completed the work to add it back.