fix: unblock upgrades
The provider tracker had an unfortunate circular dependency on the
upgrade database. This was known at the time of creation of the
provider tracker. The only way to push this through was to prevent
any changes to the underlying data for a provider tracker during
an upgrade process. This included the following data:

 1. read-only model in the model database
 2. model config in the model database
 3. cloud data in the controller database
 4. cloud credential data in the controller database

None of that information should change during a patch/build. Ideally
none of the DDL should change at all, it should always be additive.
There isn't a way to enforce that easily. For now I've added
documentation and code comments to improve the messaging to other
developers.