fix: move ifDatabaseUpgradeComplete to child workers
Having ifDatabaseUpgradeComplete on the service factory and provider
tracker prevented the upgrade. The provider tracker was waiting for
the upgrade, but couldn't do that because a child worker did the
upgrade. This caused a cycle in the dependency engine, one that was
unresolvable. This was unfortunately determined at runtime, by
releasing the gate.

The solution is to push the ifDatabaseUpgradeComplete on all the
service factory dependencies, minus the upgrade database workers
(including the gate and flags). Thus the service factory can record
the process and the upgrade database worker can control/manage
the workflow.

The API server doesn't need a ifDatabaseUpgradeComplete on it directly,
as the httpServerArgsName uses the service factory which will then
block both httpServerName and apiServerName (reducing complexity).