feat: charm domain service
Adds the scaffolding for the charm domain service. Originally, the
service package was going to offer `GetCharm` method, but that would
require serialising and deserialising the charm metadata (meta,
manifest, actions and config) for every request. When analyising the
usage patterns of the requests, it can be boiled down to wanting only
parts of the charm.

The main access pattern will then become, get the charm UUID by the
natural key. This will use the name at the start, revision and
architecture will be added later. Using the charm UUID will then
give access to a specific revision. That has a meaning, it's a unique
key to a specific charm.

Utilising the charm UUID, you can then start to ask questions about
a given charm:

 - Is it a controller charm via IsControllerCharm
 - Is it a subordinate charm via IsSubordinateCharm
 - Is it a kubernetes charm via IsKubernetesCharm

To access just the metadata you can use `GetCharmMetadata`. This
will return internal/charm.Meta. All of the API server and internal
packages use this. Internally the service can model the charm
metadata differently to satisify the core concept.

 - RunAs/CharmUser is no longer just an empty string.
 - Assumes is stored purely as a JSON blob. There is no reason to
   store/manage this in it's raw form.