feat(cli): prevent invalid controller name
The PR https://github.com/juju/juju/pull/11074 introduced the controller-name field as a possible
one to be updated through controller-config command.
It allows empty string as valid value, which is unexpected.

After this PR, empty value for this field will be rejected, as long as any value which doesn't
respect RFC 1123 for labels names (should be a string of hyphen, lowercase or digits,
not starting with an hyphen)

Moreover, I added this field to the help command.