refactor(apiserver): agenttools facade using services
The agenttools facade now uses a model config service to access model config,
and a model agent service to access the current agent version.

BREAKING CHANGE: agenttools now checks all preferred streams in preference
order, starting with the user-specified stream in `agent-stream`. Previously,
it always checked the "released" stream first, even if a custom `agent-stream`
was set, and it would only check the user-defined stream if nothing was found
in "released".

Add a new env variable to the Makefile (`JUJU_PUBLISH_STREAM`) that allows
`make simplestreams` to publish to a specified stream.

General modernising and code improvement on the agenttools facade, such as
moving internal functions to methods on the API struct.