feat: Strong type to represent machine id's.
This commit introduces a new type in the core package for representing
machine id's. At the moment we seem to be using a combination of strings
and uuid.

This follows a similar convention set up in other package to have a
stronger type that conveys meaning to the user and gives an easy
validation point for the domain layer.

My specific use case for introducing this is for a id type for the key
updater so authorised keys can be fetched for a machine. At the moment
this logic is ran off of tags but only applicable to machines. As part
of this work I would like to change the story from keys for this entity
to keys for this machine id.