test: fix run_spaces_manual_aws missing security group rule
Mongo on a manual bootstrapped controller on AWS using the
elastic public IPv4 address for the replica set host field
results in mongo needing to dial itself via the public ip.

The solution here unfortunately is to open 37017 to the world
for run_spaces_manual_aws to pass. Since this is not the
component of Juju here at test, this is an appropriate
workaround.