feat: charm manifest
Renames charm uuid to charm ID. We don't want to expose the underlying
representation of what a charm ID is. It could be a UUID it could be
a ULID. We don't care, just talk about an ID.

The service now supports getting a manifest file. Manifest files are
out of the control of juju (they're driven by the charmcraft team),
we want a nice internal representation of those files. Considering that
charms will be specific to a singular channel (and architecture), we
can keep the internal representation the same, and only change the
wire protocol. The data is correct.