fix(openstack): fix model sec group misconfiguration
When a machine fails to provision, the openstack provider attempted to
delete all security groups attached to that instance. Including any
model-scoped security groups (and potentially the openstack default sec
group!)

This can lead to errors. The firewaller configrues the model sec group
only when:
1) Once when the first machine is provioned
2) when a change to ssh-allow is detected

This means if the first machine you attempt to provision fails, the
model sec group will be deleted after it's configured. So it won't
recievd it's initial config

Attempting to deleting all secgroups attached to a machine could also
lead to countless other unexpected behaviours.

Fix this by only deleting the machine sec group when a machine fails to
provision i.e. the sec group unique to the relevant machine. This means
we will no longer attempt to delete sec groups potentially attached to
other instances.